feat: Define MPv3 to MPv3Lightweight convert CLI, updates from develop
