+++ mapping_suite_sdk/tools/entrypoints/cli/convert.py
@@ -1,29 +1,27 @@
+import json
 import logging
 from enum import Enum
 from pathlib import Path
 
 import typer
+from pydantic import ValidationError
 
 from mapping_suite_sdk import mssdk_config
+
+from mapping_suite_sdk.mapping_package_v2.adapters.mp_v2_loader import MappingPackageV2Loader
+from mapping_suite_sdk.mapping_package_v2.services.load_mapping_package_v2 import load_mapping_package_v2_from_folder
+from mapping_suite_sdk.mapping_package_v3.adapters.package_serialiser import MappingPackageV3Serialiser
+from mapping_suite_sdk.mapping_package_v3.models.mapping_package_v3_metadata_jsonld import \
+    MappingPackageV3MetadataJSONLD
 from mapping_suite_sdk.tools.entrypoints.cli import typer_verbose_callback
 from mapping_suite_sdk.tools.services.convert_mapping_package_v3 import convert_mpv3_from_mpv2
 
 logger = logging.getLogger(__name__)
 
+Version = Enum('Version', [('V2', 'v2'), ('V3', 'v3')])
 
-class BaseVersion(str, Enum):
-    V2 = "v2"
-
-
-class NewVersion(str, Enum):
-    V3 = "v3"
-
-
-BASE_VERSIONS = [BaseVersion.V2]
-NEW_VERSIONS = [NewVersion.V3]
-
-BASE_VERSION_DEFAULT = BaseVersion.V2
-NEW_VERSION_DEFAULT = NewVersion.V3
+V2 = Version.V2.value
+V3 = Version.V3.value
 
 
 def _is_already_converted(mapping_package_folder_path: Path, to_version: str) -> bool:
@@ -34,7 +32,7 @@ def _is_already_converted(mapping_package_folder_path: Path, to_version: str) ->
     possible_paths = [mapping_package_folder_path]
     if root_folder.exists():
         possible_paths.append(root_folder)
-    
+
     metadata_file = None
     for path in possible_paths:
         # Check for metadata.json or metadata.jsonld
@@ -46,36 +44,34 @@ def _is_already_converted(mapping_package_folder_path: Path, to_version: str) ->
         if test_file.exists():
             metadata_file = test_file
             break
-    
+
     if not metadata_file:
         return False
-    
-    if to_version == NewVersion.V3.value:
+
+    if to_version == V3:
         # Try to validate as V3 - if it works, it's already converted
+        metadata_dict = json.loads(metadata_file.read_text())
+        # Add path field if missing (serialiser excludes it, but model requires it)
+        if 'path' not in metadata_dict:
+            # Calculate relative path from the package folder to the metadata file
+            # Handle both direct and nested structures
+            relative_path = metadata_file.relative_to(mapping_package_folder_path)
+            metadata_dict['path'] = relative_path.as_posix()
+        # Validate by attempting to create the model
+        # If ValidationError: package is not in target version (not converted) - return False
+        # Other hard fail
         try:
-            from mapping_suite_sdk.mapping_package_v3.models.mapping_package_v3_metadata_jsonld import MappingPackageV3MetadataJSONLD
-            import json
-            metadata_dict = json.loads(metadata_file.read_text())
-            # Add path field if missing (serialiser excludes it, but model requires it)
-            if 'path' not in metadata_dict:
-                # Calculate relative path from the package folder to the metadata file
-                # Handle both direct and nested structures
-                relative_path = metadata_file.relative_to(mapping_package_folder_path)
-                metadata_dict['path'] = relative_path.as_posix()
-            # Validate by attempting to create the model
             MappingPackageV3MetadataJSONLD.model_validate(metadata_dict)
             return True
-        except Exception:
+        except ValidationError:
+            # Package is not in V3 format, so it's not converted
             return False
     return False
 
 
 def _load_mapping_package_from_folder(from_version: str, mapping_package_folder_path: Path):
     """Dynamically load a mapping package based on version."""
-    if from_version == BaseVersion.V2.value:
-        from mapping_suite_sdk.mapping_package_v2.adapters.mp_v2_loader import MappingPackageV2Loader
-        from mapping_suite_sdk.mapping_package_v2.services.load_mapping_package_v2 import load_mapping_package_v2_from_folder
-        
+    if from_version == V2:
         loader = MappingPackageV2Loader()
         return load_mapping_package_v2_from_folder(
             mapping_package_folder_path=mapping_package_folder_path,
@@ -87,7 +83,7 @@ def _load_mapping_package_from_folder(from_version: str, mapping_package_folder_
 
 def _convert_mapping_package(from_version: str, to_version: str, source_package):
     """Convert mapping package using service layer."""
-    if from_version == BaseVersion.V2.value and to_version == NewVersion.V3.value:
+    if from_version == V2 and to_version == V3:
         return convert_mpv3_from_mpv2(source_package)
     else:
         raise typer.BadParameter(f"Unsupported conversion: {from_version} -> {to_version}")
@@ -95,9 +91,7 @@ def _convert_mapping_package(from_version: str, to_version: str, source_package)
 
 def _serialise_mapping_package(to_version: str, mapping_package_folder_path: Path, converted_package):
     """Dynamically serialize a mapping package based on version."""
-    if to_version == NewVersion.V3.value:
-        from mapping_suite_sdk.mapping_package_v3.adapters.package_serialiser import MappingPackageV3Serialiser
-        
+    if to_version == V3:
         serialiser = MappingPackageV3Serialiser()
         serialiser.serialise(mapping_package_folder_path, converted_package)
     else:
@@ -110,26 +104,27 @@ def _convert_package_from_folder(from_version: str, to_version: str, mapping_pac
     converted_package = _convert_mapping_package(from_version, to_version, source_package)
     _serialise_mapping_package(to_version, mapping_package_folder_path, converted_package)
 
+
 mssdk_cli_convert_subcommand = typer.Typer(**mssdk_config.MSSDK_TYPER_DEFAULT_ARGS,
-                                          name="convert",
-                                          help="Mapping Package Conversion commands.")
+                                           name="convert",
+                                           help="Mapping Package Conversion commands.")
 
 
 @mssdk_cli_convert_subcommand.callback()
 def convert_common(
         ctx: typer.Context,
-        to_version: str = typer.Option(..., "--to-version", help=f"Target mapping package version ({NEW_VERSION_DEFAULT.value})"),
-        from_version: str = typer.Option(..., "--from-version", help=f"Source mapping package version ({BASE_VERSION_DEFAULT.value})"),
+        to_version: str = typer.Option(..., "--to-version", help=f"Target mapping package version ({V3})"),
+        from_version: str = typer.Option(..., "--from-version", help=f"Source mapping package version ({V2})"),
         verbose: bool = typer.Option(False, "--verbose", "-v",
                                      is_eager=True,
                                      callback=typer_verbose_callback),
 ) -> None:
     """Set conversion version context."""
-    if to_version not in NEW_VERSIONS:
-        raise typer.BadParameter(f"Target version must be one of {[v.value for v in NEW_VERSIONS]}, got: {to_version}")
+    if to_version != V3:
+        raise typer.BadParameter(f"Target version must be {V3}, got: {to_version}")
 
-    if from_version not in BASE_VERSIONS:
-        raise typer.BadParameter(f"Source version must be one of {[v.value for v in BASE_VERSIONS]}, got: {from_version}")
+    if from_version != V2:
+        raise typer.BadParameter(f"Source version must be {V2}, got: {from_version}")
 
     ctx.ensure_object(dict)
     ctx.obj['to_version'] = to_version
@@ -137,8 +132,8 @@ def convert_common(
 
 
 @mssdk_cli_convert_subcommand.command(**mssdk_config.MSSDK_TYPER_COMMANDS_DEFAULT_ARGS,
-                                     name="from-package",
-                                     help="Convert mapping package from package directory.")
+                                      name="from-package",
+                                      help="Convert mapping package from package directory.")
 def mssdk_cli_convert_mapping_package_from_package(
         ctx: typer.Context,
         mapping_package_path: Path = typer.Argument(..., exists=True),
@@ -169,8 +164,8 @@ def mssdk_cli_convert_mapping_package_from_package(
 
 
 @mssdk_cli_convert_subcommand.command(**mssdk_config.MSSDK_TYPER_COMMANDS_DEFAULT_ARGS,
-                                     name="from-folder",
-                                     help="Convert mapping packages from folder.")
+                                      name="from-folder",
+                                      help="Convert mapping packages from folder.")
 def mssdk_cli_convert_mapping_packages_from_folder(
         ctx: typer.Context,
         folder_path: Path = typer.Argument(...),
@@ -188,9 +183,7 @@ def mssdk_cli_convert_mapping_packages_from_folder(
     if not folder_path.is_dir():
         raise typer.BadParameter(f"Folder path is not a directory: {folder_path}")
 
-    all_converted = True
     converted_count = 0
-    failed_count = 0
 
     for mp_folder in folder_path.iterdir():
         if not mp_folder.is_dir():
@@ -203,25 +196,15 @@ def mssdk_cli_convert_mapping_packages_from_folder(
                 message=f"Package is already {to_version}, skipping conversion"))
             continue
 
-        try:
-            _convert_package_from_folder(from_version, to_version, mp_folder)
+        # Convert package - hard fail
+        _convert_package_from_folder(from_version, to_version, mp_folder)
 
-            converted_count += 1
-            logger.info(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
-                package_source=mp_folder,
-                message=f"✅ Converted {from_version} package to {to_version} package"))
-        except Exception as conversion_exception:
-            failed_count += 1
-            all_converted = False
-            logger.error(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
-                package_source=mp_folder,
-                message=f"Cannot convert mapping package: {conversion_exception}"))
+        converted_count += 1
+        logger.info(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
+            package_source=mp_folder,
+            message=f"✅ Converted {from_version} package to {to_version} package"))
 
-    status = "✅ All converted" if all_converted else "❌ Some packages failed to convert"
     logger.info(mssdk_config.MSSDK_LOGGING_MESSAGE_FORMAT.format(
         package_source=folder_path,
-        message=f"{status} ({converted_count} converted, {failed_count} failed)"))
-
-    if not all_converted:
-        raise typer.Exit(code=1)
+        message=f"✅ All converted ({converted_count} converted)"))
