+++ tests/unit/tools/entrypoints/cli/test_mssdk_cli.py
